+++ android/app/src/main/java/com/brekeke/phonedev/BrekekeMessagingService.java
@@ -11,7 +11,6 @@ import org.json.JSONArray;
 
 public class BrekekeMessagingService extends FcmInstanceIdListenerService {
   private static String TAG = "BrekekeMessagingService";
-  private static boolean alreadyGetInitialNotifications = false;
   private static ArrayList<String> initialNotifications = null;
 
   public static void getInitialNotifications(Promise promise) {
@@ -33,6 +32,10 @@ public class BrekekeMessagingService extends FcmInstanceIdListenerService {
   @Override
   public void onMessageReceived(RemoteMessage remoteMessage) {
     if (!BrekekeUtils.checkNotificationPermission(this)) {
+      return;
+    }
+    if (!BrekekeUtils.checkReadPhonePermission(this)) {
+      BrekekeUtils.emit("phonePermission", "");
       return;
     }
     BrekekeUtils.onFcmMessageReceived(this, remoteMessage.getData());