+++ src/components/navigationConfig2.ts
@@ -82,7 +82,7 @@ export const getTabs = (tab: string) => {
   const m = arr.find(_ => _.key === tab)
   if (!m) {
     RnAlert.error({
-      unexpectedErr: new Error(`Can not find sub menus for ${tab}`),
+      unexpectedErr: new Error(intl`Can not find sub menus for ${tab}`),
     })
     return []
   }