+++ src/utils/PushNotification-parse.ts
@@ -8,6 +8,7 @@ import { getCallStore } from '../stores/callStore'
 import { chatStore } from '../stores/chatStore'
 import { Nav } from '../stores/Nav'
 import { openLinkSafely, urls } from './deeplink'
+import { PushNotification } from './PushNotification'
 import { BrekekeUtils } from './RnNativeModules'
 import { toBoolean } from './string'
 import { waitTimeout } from './waitTimeout'
@@ -237,8 +238,12 @@ export const parse = async (
     console.log(
       'SIP PN debug: PushNotification-parse: local notification missed call',
     )
-    if (as.phoneappliEnabled()) {
+
+    if (as.userExtensionProperties && as.phoneappliEnabled()) {
       navIndex('goToPageCallKeypad')
+      if (Platform.OS === 'ios') {
+        PushNotification.resetBadgeNumber()
+      }
       openLinkSafely(urls.phoneappli.HISTORY_CALLED)
     } else {
       navIndex('goToPageCallRecents')