+++ src/utils/callkeep.ts
@@ -1,4 +1,10 @@
-import { AppState, Keyboard, NativeEventEmitter, Platform } from 'react-native'
+import {
+  AppState,
+  Keyboard,
+  NativeEventEmitter,
+  Platform,
+  ToastAndroid,
+} from 'react-native'
 import type { EventsPayload } from 'react-native-callkeep'
 import RNCallKeep from 'react-native-callkeep'
 
@@ -243,17 +249,20 @@ export const setupCallKeepEvents = async () => {
 
   // events from our custom BrekekeUtils module
   const eventEmitter = new NativeEventEmitter(BrekekeUtils)
-  eventEmitter.addListener('answerCall', (uuid: string) => {
+  eventEmitter.addListener('answerCall', async (uuid: string) => {
     // should update the native android UI here to fix a case with auto answer
     const c = cs.calls.find(_ => _.callkeepUuid === uuid && _.answered)
     if (c) {
+      if (Platform.OS === 'android') {
+        // with auto answer, talkingAvatar takes too long to update
+        BrekekeUtils.setTalkingAvatar(
+          uuid,
+          c.talkingImageUrl,
+          c.partyImageSize === 'large',
+        )
+        await waitTimeout(17)
+      }
       BrekekeUtils.onCallConnected(uuid)
-      // with auto answer, talkingAvatar takes too long to update
-      BrekekeUtils.setTalkingAvatar(
-        uuid,
-        c.talkingImageUrl,
-        c.partyImageSize === 'large',
-      )
     }
     cs.onCallKeepAnswerCall(uuid.toUpperCase())
     RNCallKeep.setOnHold(uuid, false)
@@ -309,14 +318,27 @@ export const setupCallKeepEvents = async () => {
     }
     cs.onSelectBackgroundCall(c)
   })
-  eventEmitter.addListener('makeCall', async (phoneNumber: string) => {
-    if (!(await getAuthStore().autoSignInLast())) {
-      return
+  eventEmitter.addListener('phonePermission', () => {
+    console.log(
+      'CallKeep debug: phonePermission currentState' + AppState.currentState,
+    )
+    if (AppState.currentState === 'active') {
+      ToastAndroid.showWithGravity(
+        intl`Incoming call blocked. Please allow phone permission in settings to receive calls`,
+        ToastAndroid.LONG,
+        ToastAndroid.BOTTOM,
+      )
     }
-    await waitSip()
-    getCallStore().startCall(phoneNumber)
   })
   // other utils
+  eventEmitter.addListener('onIncomingCallActivityBackPressed', () => {
+    if (!RnStacker.stacks.length) {
+      nav.goToPageIndex()
+    } else {
+      RnStacker.stacks = [RnStacker.stacks[0]]
+    }
+    cs.inPageCallManage = undefined
+  })
   eventEmitter.addListener('onBackPressed', onBackPressed)
   eventEmitter.addListener('onIncomingCallActivityBackPressed', () => {
     if (!RnStacker.stacks.length) {